фыв

+++ Laboratory 5/RolePage.xaml.cs
@@ -31,15 +31,34 @@ namespace Laboratory_5
 
         private void Button_Click(object sender, RoutedEventArgs e)
         {
-            Role.InsertQuery(RoleBox.Text);
-            RoleGrid.ItemsSource = Role.GetData();
+            double num = 0.0;
+            if (RoleBox.Text == null || RoleBox.Text == "")
+            {
+                MessageBox.Show("Введите данные в поля");
+            }
+            else if ((double.TryParse(RoleBox.Text, out num)))
+            {
+                MessageBox.Show("Вы ввели число");
+            }
+            else
+            {
+                Role.InsertQuery(RoleBox.Text);
+                RoleGrid.ItemsSource = Role.GetData();
+            }
         }
 
         private void Button_Click_1(object sender, RoutedEventArgs e)
         {
-            int id = (int)(RoleGrid.SelectedValue as DataRowView).Row[0];
-            Role.DeleteQuery(id);
-            RoleGrid.ItemsSource = Role.GetData();
+            if (RoleGrid.SelectedItem == null)
+            {
+                MessageBox.Show("Выберите элемент для удаления");
+            }
+            else
+            {
+                int id = (int)(RoleGrid.SelectedValue as DataRowView).Row[0];
+                Role.DeleteQuery(id);
+                RoleGrid.ItemsSource = Role.GetData();
+            }
         }
 
         private void Button_Click_2(object sender, RoutedEventArgs e)
